+++ src/main/java/zd/zero/waifu/motivator/plugin/assets/RemoteAssetDefinitionService.kt
@@ -3,9 +3,9 @@ package zd.zero.waifu.motivator.plugin.assets
 import kotlin.random.Random
 
 abstract class RemoteAssetDefinitionService<T : AssetDefinition>(
-    protected val remoteAssetManager: RemoteAssetManager<T>
+    private val remoteAssetManager: RemoteAssetManager<T>
 ) {
-    protected val random = Random(System.currentTimeMillis())
+    private val random = Random(System.currentTimeMillis())
 
     fun getRandomAssetByCategory(waifuAssetCategory: WaifuAssetCategory): T =
         remoteAssetManager.resolveAsset(
